Prepnutie ADC na rozlíšenie 12 bitov (pôvodne 8)
Prevodník meria v rozsahu 12 bitov, tým pádom musí byť neustále zapnuté
premapovanie hodnôt. Nie kvôli kalibrácii, ale kvôli rozsahu hodnôt (v
konečnej fáze chceme aj tak len 8 bitov).

+++ src/lib.h
@@ -18,7 +18,7 @@ void initUSART(void);
 
 void initADCwithDMA(void);
 
-int *remap(uint8_t ADC_Values[2]);
+int *remap(uint16_t ADC_Values[2]);
 char *INTconversionCHAR(uint8_t value);
